+++ src/pages/directory/create.tsx
@@ -3,26 +3,11 @@ import Header from '@src/components/Header';
 import Sidebar from '@src/components/Sidebar';
 import ContactSelector from '@src/components/CreateContactSelector';
 import OfficerSelector from '@src/components/OfficerSelector';
-import { z } from 'zod';
+import { type z } from 'zod';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { selectContact } from '@src/server/db/models';
+import { createClubSchema } from '@src/utils/formSchemas';
 
-export const createClubSchema = z.object({
-  name: z.string().min(3),
-  description: z.string().min(1),
-  officers: z
-    .object({
-      name: z.string().min(1),
-      position: z.string().min(1),
-    })
-    .array()
-    .min(1),
-  contacts: selectContact
-    .omit({ clubId: true, url: true })
-    .extend({ url: z.string().min(1) })
-    .array(),
-});
 const Page = () => {
   const {
     register,
